funcionando

+++ app/src/main/java/com/example/flowsyroomfrancisco/data/repositories/BlogRepository.kt
@@ -1,6 +1,8 @@
 package com.example.flowsyroomfrancisco.data.repositories
 
+import android.util.Log
 import com.example.flowsyroomfrancisco.data.BlogDao
+import com.example.flowsyroomfrancisco.data.model.toBlog
 import com.example.flowsyroomfrancisco.data.model.toBlogEntity
 import com.example.flowsyroomfrancisco.data.sources.remote.RemoteDataSource
 import com.example.flowsyroomfrancisco.domain.model.Blog
@@ -15,12 +17,21 @@ class BlogRepository @Inject constructor(
     private val remoteDataSource: RemoteDataSource,
     private val blogDao: BlogDao
 ){
-    fun getAllBlogs(): Flow<NetworkResultt<List<Blog>>> {
-        return flow {
-            emit(NetworkResultt.Loading())
+    fun getAllBlogs(): Flow<NetworkResultt<List<Blog>>> = flow {
+        emit(NetworkResultt.Loading())
+
+        //Consultar la base de datos local
+        val localBlogs = blogDao.getAllBlogs()
+        if (localBlogs.isNotEmpty()) {
+            Log.d("Repository", "Loading data from local database")
+            emit(NetworkResultt.Success(localBlogs.map { it.toBlog() }))
+        } else {
+            //Realizar la solicitud al servidor solo si no hay datos locales
             val result = remoteDataSource.getAllBlogs()
             emit(result)
-            if (result is NetworkResultt.Success){
+
+            //Actualizar la base de datos local si la solicitud fue exitosa
+            if (result is NetworkResultt.Success) {
                 result.data?.let { lista ->
                     blogDao.deleteBlogList(lista.map {
                         it.toBlogEntity()
@@ -29,7 +40,8 @@ class BlogRepository @Inject constructor(
                         it.toBlogEntity()
                     })
                 }
+                Log.d("Repository", "Loading data from server")
             }
-        }.flowOn(Dispatchers.IO)
-    }
+        }
+    }.flowOn(Dispatchers.IO)
 }
